finish moving all the automation around
